Owner of the image is authorized to edit and delete image

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -94,12 +94,23 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
+        User loggeduser = (User) session.getAttribute("loggeduser");
+
+        model.addAttribute("image", image);
+
+        //Check if the user of the current image is same as that of logged-in user
+        if (!image.getUser().getId().equals(loggeduser.getId())) {
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
 
         String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
         model.addAttribute("tags", tags);
+
         return "images/edit";
     }
 
@@ -134,7 +145,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images";
     }
 
 
@@ -142,7 +153,22 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,
+                                    Model model,
+                                    HttpSession session) {
+        Image currentImage = imageService.getImage(imageId);
+        List<Tag> tags = currentImage.getTags();
+        model.addAttribute("tags", tags);
+        model.addAttribute("image", currentImage);
+
+        //If current image dont belongs to logged in user
+        User loggeduser = (User) session.getAttribute("loggeduser");
+        if (!currentImage.getUser().getId().equals(loggeduser.getId())) {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
